General-purpose circle-rectangle collision check (for backup only)
This is not really needed. We can use a much simpler check, if we assume
that the initial state is collision free.

However, I only figured that out after writing the method ;)
So, I'll commit it to keep a copy of it and I'll remove it in a
subsequent commit.